Implemen HashSum
There are a few problems here, but this is almost working. At the very
least it works on my simple cases.

First problem: the code does not seem to be thread safe as running `go
test` breaks but the examples work in isolation.

Second problem: I haven't put enough thought into testing overflow
buckets and how to ensure I get a hash collision. There are likely ideas
in the tests of the go runtime, but I haven't had a chance to dig into
those.