Get pointer to hmap directly, not from interface
I believe the way this was working before was only accidentally working.
The wrapping in an interface trick would only work reliably if the
function signature expected an interface. Instead, the casts were only
coincidentally finding the pointer to the hmap.

This is likely way the test code was failing with seg faults -- the hmap
pointer did not happen to be in that exact location in memory when
called from the test runner code.

I'm still not 100% sure why. I tried looking at the addresses in delve
but it would not let me see them. In any case, the way Elliott has done
this with the double pointer does feel like the "correct" way to do it.