[lsm-trees] Design a simple format for point reads
When the file is opened, the index section is read into memory (with the
help of the two-byte header section to find where to stop reading the
index) and stored as a map of the string key to the offset values. A
point read looks up the key in the map and then seeks to that offset in
the file.

The file format has three sections.

Section 1: Header (2 bytes)
---------------------------
A uint16 representing the length of the Index section

Section 2: Index (len(index) bytes)
-----------------------------------
These are index entries packed one after the other. Each entry has a
six-byte header followed by the key itself. The six-byte header consists
of three uint16s: the value's offset into the data section as well as the
length of the key and value strings.

Section 3: Data (len(values) bytes)
-----------------------------------
These are simply the values themselves packed one after the other.