[lsm-trees] Support range scans with a key array
This is the simplest thing I could think of, but it's not particular
memory efficient since we're storing the keys twice in memory (once in
the index, once in this array). But this approach supports the feature
without any modifications to the file format and doesn't require any
extra passes through the data structure when reading the index, so
seemed reasonable.