[roaring-bitmaps] Implement WAH compression
There's a lot of complex bit-shifting et cetera going on so I tried to
be overly verbose with the comments, just to mostly explain to myself
what was going on.

Essentially there's just a bunch of accounting going on between
reading/writing w-1 bits at a time out of an array that gives w bits a
time. You're almost always between two elements of that array and so
have to bitshift and mask to split the two parts.