[distsys-intro] Parse input and store in memory
As of now the only operations supported are `get` and `set` (case
insensitive). There is simply a global `map[string]string` that is
updated.

There is minimal valid input checking, mainly that the first word is
either get or set and that set operations are followed by 'key=value'.
These are then placed in a `Command` struct that is then read by the
function executing the command.