[distsys-intro] Split server/REPL with sockets
Using Unix domain sockets over a tmp file, which seems just fine for
now. There are two `main()` functions in the same package, but I didn't
want to overcomplicate the packaging for now. The server can be run with
`go run server.go` and the REPL with `go run repl.go`.