[communication] Use protos for client-server comms
Instead of creating a binary wire format that would have been a poor
implementation of protocol buffers, just use them. This also gives
practice with using a real thing.

This required a bit of a project reorganization and the message
themselves are really simple but the nice thing is that it can evolve as
the project goes.